+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Added check if order was paid by gifcard, then will change payment method to one of giftcard payment methods.
 - Added translations for some checkout fields. (Thanks to @Davie82)
 
 ### Fixed